Align code examples
The tutorial text assigns most of the code examples to variables, then sometimes doesn't explicitly display the results in the text.
For the notebook, I've updated the code so the code snippets assign to variable like they do in the tutorial, and then added calls to .head() and sometimes also displayed .shape in order to visualize the results for readers in the notebook.

There were also a few inconsistencies in regards to using rsuffix as an additional argument to lsuffix, which I removed from the notebook to align with the tutorial text.